Document differences between double logistic scripts
